+++ R/paired.R
@@ -85,6 +85,14 @@ paired <- function(formula, data, id, na.action, subset=NULL, control = NULL, ..
     by.col <- droplevels(by.col)
     by.levels <- levels(by.col)
   } else by.levels <- sort(unique(by.col))
+  by.col <- as.character(by.col)
+  by.levels <- as.character(by.levels)
+  if(any(by.levels == ""))
+  {
+    warning('Empty string detected in time point is not allowed; converting to " ".')
+    by.col[by.col == ""] <- " "
+    by.levels <- unique(replace(by.levels, by.levels == "", " "))
+  }
 
   if(length(by.levels) != 2) stop("Please specify exactly 2 time points")
   ids <- modeldf$`(id)`
@@ -202,7 +210,7 @@ paired <- function(formula, data, id, na.action, subset=NULL, control = NULL, ..
       {
         for(bylev in by.levels) {
           idx <- by.col == bylev
-          bystatlist[[as.character(bylev)]] <- do.call(statfun, list(currcol[idx], levels=xlevels, na.rm=TRUE, ...))
+          bystatlist[[bylev]] <- do.call(statfun, list(currcol[idx], levels=xlevels, na.rm=TRUE, ...))
         }
       }
       if(statfun %in% c("countpct", "countrowpct", "countcellpct"))