+++ tests/testthat/test_tableby.R
@@ -824,3 +824,6 @@ test_that("06/19/2018: term.name (#109)", {
     )
   )
 })
+
+mockstudy$grp <- c(rep("Group1", 749), rep("Group2",749), "")
+test_that("08/23/2018: empty string in by-variable (#121)", expect_warning(summary(tableby(grp ~ race, data=mockstudy)), "Empty"))